+++ src/components/InterfaceScore/index.scss
@@ -90,7 +90,7 @@
   }
 
   &__animation {
-    width: 100%;
+    width: 300px;
   }
 
   &__sign {
