+++ src/components/MotionShared/index.scss
@@ -6,11 +6,6 @@
   height: 100%;
   background: $black;
 
-  &.hdie {
-    visibility: hidden;
-    pointer-events: none;
-  }
-
   &__overlay {
     width: 100%;
     height: 100%;
